Time: 3 ms (94.62%), Space: 40.4 MB (36.6%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean judgeSquareSum(int c) {
         long left = 0;
-        long right = (int) Math.sqrt(c);
+        long right = (long) Math.sqrt(c);
         
         while (left <= right) {
             long sum = left * left + right * right;
